blkzone: fix whole device detection
The current detection code for chunk size assumes that the
underlying device is a device that uses the minor device id
as the partition id, and that the whole device has minor id 0.
This is not true for example null_blk and nvme device drivers.

Therefore, use blkid's blkid_probe_get_wholedisk_devno() to
resolve the whole device minor id appropriately.

Signed-off-by: Matias Bjørling <matias.bjorling@wdc.com>
Tested-by: Damien Le Moal <Damien.lemoal@wdc.com>