su: set SETGID bit
Without the bit set only the effective group ID is changed but not
the real group id. (If the shell is bash, the effective group id
gets reset to the unchanged real group ID, and no startup file is
read. See "Invoked with unequal effective and real UID/GIDs" under
"6.2 Bash Startup Files" of the bash manual.)