changed: util-linux/libmount/src/tab.c
Added mnt_table_move_safe()
Added mnt_table_insert_safe()
Added mnt_table_remove_safe()
The functions are checking for table memberships of the respective entries if you are not sure about it. So you can use the "unsafe" versions if you know that erverything is alright and avoid the overhead of checking the memberships
I don't know if this is an appropriate alternative but this avoids looping over the table over and over again, although you know that every entry belongs to the right table before calling the function.